Ajustando configurações e atualizando a conteinerização
