--- app/projeto_robotica.egg-info/PKG-INFO
@@ -1,14 +1,0 @@
-Metadata-Version: 2.1
-Name: projeto_robotica
-Version: 0.1.0
-Summary: Repositório para guardar os códigos do projeto de robótica (braço mecanico)
-Author: IanKarlo, email= <iankarlots@gmail.com>
-Author-email: EGmux <egb2@cin.ufpe.br>, Ek <ek@gmail.com>
-License: MIT
-Description-Content-Type: text/markdown
-Requires-Dist: matplotlib
-Requires-Dist: numpy
-Requires-Dist: roboticstoolbox-python
-Requires-Dist: scipy
-Requires-Dist: basedpyright
-Requires-Dist: flake8
