--- app/projeto_robotica.egg-info/requires.txt
@@ -1,6 +1,0 @@
-matplotlib
-numpy
-roboticstoolbox-python
-scipy
-basedpyright
-flake8
