--- app/projeto_robotica.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-projeto_robotica
