--- app/pyproject.toml
@@ -1,25 +1,0 @@
-[project]
-name = "projeto_robotica"
-version = "0.1.0"
-description = "Repositório para guardar os códigos do projeto de robótica (braço mecanico)"
-authors = [{name="IanKarlo, email= <iankarlots@gmail.com>"},{name="EGmux", email= "egb2@cin.ufpe.br"},{name = "Ek", email= "ek@gmail.com"}]
-license = {text="MIT"}
-readme = "README.md"
-dependencies = [
-	"matplotlib",
-	"numpy",
-	"roboticstoolbox-python",
-	"scipy",
-	"basedpyright",
-	"flake8"
-]
-
-[build-system]
-requires = ["setuptools"]
-
-
-[project.scripts]
-first_question = "projeto_robotica.tasks:test_first_question"
-second_question = "projeto_robotica.tasks:test_second_question"
-start = "projeto_robotica.tasks:start"
-
