+++ projeto_robotica/functions/one_dim.py
