+++ projeto_robotica/functions/trajectory/motor.py
