+++ projeto_robotica/temp.py
