+++ requirements.txt
@@ -1,6 +1,3 @@
-matplotlib
-numpy
-basedpyright
-roboticstoolbox-python
-flake8
-scipy
+matplotlib==3.5.3
+numpy==1.21.6
+roboticstoolbox-python==1.1.1
